second commit

+++ src/app/components/signup/signup.component.ts
@@ -1,7 +1,8 @@
-import { Component } from '@angular/core';
+import { Component , OnInit} from '@angular/core';
 import { FormControl, FormGroup, Validators } from '@angular/forms';
 import { Router } from '@angular/router';
-import { AuthService } from 'src/app/services/auth.service';
+import { Users } from './../../interfaces/users';
+
 
 
 
@@ -11,66 +12,61 @@ import { AuthService } from 'src/app/services/auth.service';
   templateUrl: './signup.component.html',
   styleUrls: ['./signup.component.css']
 })
-export class SignupComponent {
-  users:any[]=[];
-  request:string="";
-  guest_session_id:string=""
+export class SignupComponent implements OnInit{
+  users:Users[]=[];
+  SignUpError:boolean=false;
 
   
-  constructor(private _AuthService:AuthService , private _Router:Router){
-    this.GetAllUsers()
+  constructor(private _Router:Router){
+
+  }
+  
+  ngOnInit():void{
+    if (localStorage.getItem('users') !== null) {
+      this.users = JSON.parse(localStorage.getItem('users')!); 
+      console.log(this.users);
+    }
 
   }
 
-GetAllUsers(){
-  if (localStorage.getItem('users') !== null) {
-    this.users = JSON.parse(localStorage.getItem('users')!); // Parse the JSON string into an object or array
-    console.log(this.users);
-  }
-}
-
-CreateRequestToken(){
-  return this._AuthService.RequestToken().subscribe((res)=>{
-   this.request=res.request_token;
-   localStorage.setItem('request',this.request)
-   console.log(localStorage.getItem('request'))
-   window.location.href=`https://www.themoviedb.org/authenticate/${this.request}?redirect_to=https://movie-app-de9z-esy4jtjli-alaa-mostafaas-projects.vercel.app/login`
 
 
-  })
-}
 
-// Sign Up Form
+
+// Sign Up Form and its validations
 
 SignUpForm:FormGroup=new FormGroup({
-  name:new FormControl(null,[Validators.required]),
+  name:new FormControl(null,[Validators.required,Validators.minLength(3),Validators.maxLength(15)]),
   email:new FormControl(null,[Validators.required,Validators.email]),
-  password:new FormControl(null,[Validators.required]),
-  session_id:new FormControl(null)
+  password:new FormControl(null,[Validators.required,Validators.minLength(4)]),
 })
 
 
+// Sign Up function
 
-   SignUp(form:FormGroup){
-  console.log(form.value)
+SignUp(form:FormGroup){
+  let isExist;
+
+  for(let user of this.users){
+    if( user.email === form.value.email){
+      isExist=true;
+    }else{
+      isExist=false
+    }
+  }
+
+if(isExist){
+  this.SignUpError=true
+
+}else{
+  this.SignUpError=false
   this.users.push(form.value)
   localStorage.setItem('users',JSON.stringify(this.users))
-  this.GetAllUsers()
-  this.CreateRequestToken()
-
-
-
+  this._Router.navigate(['/login'])
 }
 
 
-
-
-
-/* 
-       */
-
-
-
+}
